refactor(list): convert MatLineSetter into a function
Currently the `MatLineSetter` is a class, however it doesn't really do anything aside from calling a couple of functions inside the constructor. The following changes turn it into a function and make it a bit more compact. This also avoids having to turn off a tslint rule in a couple of places.